technical-notes/testing-projects/load-testing-selenium/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
requests==2.31.0
locust==2.17.0
faker==20.1.0
python-dotenv==1.0.0
allure-pytest==2.13.2
pytest-benchmark==4.0.0
pytest-timeout==2.1.0
pytest-repeat==0.9.1
pytest-parallel==0.1.1
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.3 